third_party/WebKit/LayoutTests/editing/inserting/space-after-removeformat.html

<!doctype HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
// This is a test for http://wkb.ug/14062
test(() => assert_selection(
    '<div contenteditable>^<span>Some text.</span>|</div>',
    'removeFormat',
    '<div contenteditable><span>^Some text.|</span></div>'),
    'RemoveFormat command should not convert U+0020 to U+00A0');
</script>
